FIA-76 - moved all models from common.api.accounts and common.account
