--- src/fianchetto_tradebot/common/account/etrade/etrade_account.py
@@ -1,5 +1,0 @@
-from fianchetto_tradebot.common.account.account import Account
-
-
-class ETradeAccount(Account):
-    account_id_key: str
