+++ src/fianchetto_tradebot/common/api/accounts/account_service.py
@@ -1,10 +1,11 @@
-from fianchetto_tradebot.common.api.accounts.list_accounts_response import ListAccountsResponse
-from fianchetto_tradebot.common.api.accounts.get_account_balance_request import GetAccountBalanceRequest
-from fianchetto_tradebot.common.api.accounts.get_account_balance_response import GetAccountBalanceResponse
-from fianchetto_tradebot.common.api.accounts.get_account_request import GetAccountRequest
-from fianchetto_tradebot.common.api.accounts.get_account_response import GetAccountResponse
+
 from fianchetto_tradebot.common.api.api_service import ApiService
 from fianchetto_tradebot.common.brokerage.connector import Connector
+from fianchetto_tradebot.common_models.api.account.get_account_balance_request import GetAccountBalanceRequest
+from fianchetto_tradebot.common_models.api.account.get_account_balance_response import GetAccountBalanceResponse
+from fianchetto_tradebot.common_models.api.account.get_account_request import GetAccountRequest
+from fianchetto_tradebot.common_models.api.account.get_account_response import GetAccountResponse
+from fianchetto_tradebot.common_models.api.account.list_accounts_response import ListAccountsResponse
 
 
 class AccountService(ApiService):
