+++ src/fianchetto_tradebot/common_models/account/etrade/etrade_account.py
@@ -1,0 +1,5 @@
+from fianchetto_tradebot.common_models.account.account import Account
+
+
+class ETradeAccount(Account):
+    account_id_key: str
